Many logic in QuizEntityForm class, split form definition to helper class.
